Consolidate project into single main binary
