+++ Cargo.toml
@@ -12,12 +12,3 @@ hex = "0.4"
 chrono = "0.4"
 rand = "0.8"
 data-encoding = "2.*"
-
-
-[[bin]]
-name = "syllable_frequency"
-path = "src/bin/syllable_frequency.rs"
-
-[[bin]]
-name = "humanise"
-path = "src/bin/humanise.rs"
